Adjust android part.
